refactor: strategy location
