+++ src/test/java/LottoMachineTest.java
@@ -1,4 +1,5 @@
 import domain.*;
+import domain.strategy.LottoNumberStrategy;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
